lifted interfaces to root file, deleted interface component files
