--- src/components/Description/Description.interface.ts
@@ -1,4 +1,0 @@
-    export interface DescriptionProps {
-    description: string | null;
-    iconId: string | null;
-    }
