--- src/components/ErrorDisplay/ErrorDisplay.interface.ts
@@ -1,3 +1,0 @@
-export interface ErrorDisplayProps {
-    cityId:string;
-}
