+++ src/components/MainIcon/index.tsx
