--- src/components/SunriseSunset/SunriseSunset.interface.ts
@@ -1,5 +1,0 @@
-export interface SunriseSunsetProps {
-    timezone: number,
-    sunrise: number,
-    sunset: number
-}
