--- src/components/temperatures/Temperatures.interaces.ts
@@ -1,11 +1,0 @@
-export interface TemperaturesProps {
-  temps: {
-    temp: number;
-    tempMin: number;
-    tempMax: number;
-    feelsLike: number;
-    humidity: number;
-    pressure: number;
-  };
-  currentUnits?: 'Celcius' | 'Fahrenheit';
-}
