--- src/components/unitsButton/UnitButton.interfaces.ts
@@ -1,4 +1,0 @@
-export interface UnitButtonProps {
-    handleChangeUnits: () => void;
-    currentUnits: 'Celcius' | 'Fahrenheit';
-}
