--- src/components/wind/Wind.interfaces.ts
@@ -1,9 +1,0 @@
-import { UnitButtonProps } from '../UnitsButton/UnitButton.interfaces';
-
-export interface WindProps {
-  wind: {
-    speed: number;
-    deg: number;
-  };
-  currentUnits?: 'Celcius' | 'Fahrenheit';
-}
